More Fixes

+++ Resources/Services/IAction.cs
@@ -41,23 +41,35 @@ namespace DSManager.Resources.Services
     {
         private readonly DataModel _data;
         private readonly ObservableCollection<DataModel> _collection;
+        private readonly int _originalIndex;
 
         public DeleteAction(DataModel data, ObservableCollection<DataModel> collection)
         {
             _data = data;
             _collection = collection;
+            _originalIndex = data.Id;
         }
-
+        //Пофиксить эту шляпу
         public void Undo()
         {
-            _collection.Add(_data);
-            ExcelService.DeleteRowFromExcelFile(_data.Id, 0);
+            if (_originalIndex >= 0 && _originalIndex < _collection.Count)
+            {
+                _collection.Insert(_originalIndex, _data);
+                ExcelService.InsertRow(_originalIndex, _data.FIO, _data.Department, _data.Setup, _data.Start, _data.End, _data.Status);
+            }
+            else
+            {
+                _collection.Add(_data);
+                ExcelService.AddRow(_data.FIO, _data.Department, _data.Setup, _data.Start, _data.End, _data.Status);
+            }
         }
-
         public void Redo()
         {
-            _collection.Remove(_data);
-            ExcelService.AddRow(_data.FIO, _data.Department, _data.Setup, _data.Start, _data.End, _data.Status);
+            if (_collection.Contains(_data))
+            {
+                _collection.Remove(_data);
+                ExcelService.DeleteRowFromExcelFile(_data.Id, 0);
+            }
         }
     }
 }
